Move ApplicationMapService bookmark logic into its own service
