+++ projects/core/src/lib/map/services/application-map.service.ts
@@ -1,51 +1,27 @@
 import { Injectable, OnDestroy } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { LayerModel, LayerTypesEnum, MapService, OgcHelper, WMSLayerModel, WMTSLayerModel } from '@tailormap-viewer/map';
-import { combineLatest, concatMap, distinctUntilChanged, filter, forkJoin, map, Observable, of, Subject, take, takeUntil, tap, distinctUntilKeyChanged, debounceTime } from 'rxjs';
+import { combineLatest, concatMap, distinctUntilChanged, filter, forkJoin, map, Observable, of, Subject, take, takeUntil, tap } from 'rxjs';
 import { ResolvedServerType, ServiceModel, ServiceProtocol } from '@tailormap-viewer/api';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { ArrayHelper } from '@tailormap-viewer/shared';
-import { selectLoadStatus, selectMapOptions, selectOrderedVisibleBackgroundLayers,
-  selectOrderedVisibleLayersWithServices, selectLayers, selectLayerTreeNodes } from '../state/map.selectors';
+import { selectMapOptions, selectOrderedVisibleBackgroundLayers, selectOrderedVisibleLayersWithServices } from '../state/map.selectors';
 import { ExtendedAppLayerModel } from '../models';
 import { selectCQLFilters } from '../../filter/state/filter.selectors';
-import { LoadingStateEnum } from '@tailormap-viewer/shared';
-import { BookmarkFragmentStringDescriptor, BookmarkFragmentProtoDescriptor } from '../../bookmark/bookmark.models';
-import { BookmarkService } from '../../bookmark/bookmark.service';
-import { setLayerVisibility, setLayerOpacity, setLayerTreeNodeChildren } from '../state/map.actions';
-import { LayerVisibilityBookmarkFragment, LayerTreeOrderBookmarkFragment } from '../bookmark/bookmark_pb';
-import { MapBookmarkHelper } from '../bookmark/bookmark.helper';
+import { ApplicationMapBookmarkService } from './application-map-bookmark.service';
 
 @Injectable({
    providedIn: 'root',
 })
 export class ApplicationMapService implements OnDestroy {
-  private static LOCATION_BOOKMARK_DESCRIPTOR: BookmarkFragmentStringDescriptor = { type: 'string', identifier: '' };
-  private static VISIBILITY_BOOKMARK_DESCRIPTOR: BookmarkFragmentProtoDescriptor<LayerVisibilityBookmarkFragment> = {
-    type: 'proto',
-    proto: LayerVisibilityBookmarkFragment,
-    identifier: '1',
-  };
-
-  private static ORDERING_BOOKMARK_DESCRIPTOR: BookmarkFragmentProtoDescriptor<LayerTreeOrderBookmarkFragment> = {
-    type: 'proto',
-    proto: LayerTreeOrderBookmarkFragment,
-    identifier: '2',
-  };
-
   private destroyed = new Subject();
   private capabilities: Map<number, string> = new Map();
-
-  // Set to true once the bookmark has been read out at least once. This
-  // is to make sure that, on slower(?) systems, the bookmark is checked
-  // before the location of the map is initially written back.
-  private bookmarkChecked = false;
 
   constructor(
     private store$: Store,
     private mapService: MapService,
     private httpClient: HttpClient,
-    private bookmarkService: BookmarkService,
+    private applicationMapBookmarkService: ApplicationMapBookmarkService,
   ) {
     const isValidLayer = (layer: LayerModel | null): layer is LayerModel => layer !== null;
     this.store$.select(selectMapOptions)
@@ -85,88 +61,6 @@ export class ApplicationMapService implements OnDestroy {
       .subscribe(([ layers, layerManager ]) => {
         layerManager.setLayers(layers.filter(isValidLayer));
       });
-
-    combineLatest([
-      this.bookmarkService.registerFragment$(ApplicationMapService.VISIBILITY_BOOKMARK_DESCRIPTOR),
-      this.store$.select(selectLayers),
-      this.store$.select(selectLoadStatus),
-    ]).pipe(
-        takeUntil(this.destroyed),
-        filter(([ ,, loadStatus ]) => loadStatus === LoadingStateEnum.LOADED),
-        distinctUntilKeyChanged('0'),
-      )
-      .subscribe(([ fragment, layers ]) => {
-        const bookmarkContents = MapBookmarkHelper.visibilityDataFromFragment(fragment, layers);
-        if (bookmarkContents.visibilityChanges.length > 0) {
-          this.store$.dispatch(setLayerVisibility({ visibility: bookmarkContents.visibilityChanges }));
-        }
-
-        for (const item of bookmarkContents.opacityChanges) {
-          this.store$.dispatch(setLayerOpacity(item));
-        }
-      });
-
-    combineLatest([
-      this.bookmarkService.registerFragment$(ApplicationMapService.ORDERING_BOOKMARK_DESCRIPTOR),
-      this.store$.select(selectLayerTreeNodes),
-      this.store$.select(selectLoadStatus),
-    ]).pipe(
-        takeUntil(this.destroyed),
-        filter(([ ,, loadStatus ]) => loadStatus === LoadingStateEnum.LOADED),
-        distinctUntilKeyChanged('0'),
-      )
-      .subscribe(([ fragment, layers ]) => {
-        const bookmarkContents = MapBookmarkHelper.layerTreeOrderFromFragment(fragment, layers);
-        for (const [ key, children ] of bookmarkContents) {
-          this.store$.dispatch(setLayerTreeNodeChildren({ nodeId: key, children }));
-        }
-      });
-
-    this.getVisibilityBookmarkData()
-      .pipe(
-        takeUntil(this.destroyed),
-      )
-      .subscribe(bookmark => {
-          this.bookmarkService.updateFragment(ApplicationMapService.VISIBILITY_BOOKMARK_DESCRIPTOR, bookmark);
-      });
-
-    this.getOrderBookmarkData()
-      .pipe(
-        takeUntil(this.destroyed),
-      )
-      .subscribe(bookmark => {
-          this.bookmarkService.updateFragment(ApplicationMapService.ORDERING_BOOKMARK_DESCRIPTOR, bookmark);
-      });
-
-    combineLatest([ this.mapService.getMapViewDetails$(), this.mapService.getUnitsOfMeasure$() ])
-      .pipe(takeUntil(this.destroyed))
-      .subscribe(([ info, measure ]) => {
-        const fragment = MapBookmarkHelper.fragmentFromLocationAndZoom(info, measure);
-        if (fragment !== undefined && this.bookmarkChecked) {
-          this.bookmarkService.updateFragment(ApplicationMapService.LOCATION_BOOKMARK_DESCRIPTOR, fragment);
-        }
-      });
-
-    combineLatest([
-      this.bookmarkService.registerFragment$(ApplicationMapService.LOCATION_BOOKMARK_DESCRIPTOR),
-      this.mapService.getMapViewDetails$(),
-      this.mapService.getUnitsOfMeasure$(),
-    ])
-      .pipe(
-        takeUntil(this.destroyed),
-        debounceTime(0),
-        distinctUntilKeyChanged('0'),
-      )
-      .subscribe(([ descriptor, viewDetails, unitsOfMeasure ]) => {
-        const centerAndZoom = MapBookmarkHelper.locationAndZoomFromFragment(descriptor, viewDetails, unitsOfMeasure);
-        this.bookmarkChecked = true;
-
-        if (centerAndZoom === undefined) {
-          return;
-        }
-
-        this.mapService.setCenterAndZoom(centerAndZoom[0], centerAndZoom[1]);
-      });
   }
 
   public selectOrderedVisibleLayersWithFilters$() {
@@ -188,31 +82,6 @@ export class ApplicationMapService implements OnDestroy {
       this.mapService.getLayerManager$().pipe(take(1)),
     ]);
   }
-
-  private getVisibilityBookmarkData() {
-    return combineLatest([
-      this.store$.select(selectLayers),
-      this.store$.select(selectLoadStatus),
-    ]).pipe(
-      filter(([ , loadStatus ]) => loadStatus === LoadingStateEnum.LOADED),
-      map(([layers]) => {
-        return MapBookmarkHelper.fragmentFromVisibilityData(layers);
-      }),
-    );
-  }
-
-  private getOrderBookmarkData() {
-    return combineLatest([
-      this.store$.select(selectLayerTreeNodes),
-      this.store$.select(selectLoadStatus),
-    ]).pipe(
-      filter(([ , loadStatus ]) => loadStatus === LoadingStateEnum.LOADED),
-      map(([layers]) => {
-        return MapBookmarkHelper.fragmentFromLayerTreeOrder(layers);
-      }),
-    );
-  }
-
 
   public ngOnDestroy() {
     this.destroyed.next(null);
